Added intro, combining intro + overview
So instead of cloning the repo and adding a sub-module to the collections directory, or adding a new card within the Proxima collection, I just chose the path of least resistance and added the introduction to this overview. I figure we can adjust this later, but the intro needs to be live by tomorrow, so I'm thinking the important thing for now is just that it's in there.